+++ FrequencyByte.h
@@ -9,13 +9,13 @@ using namespace std;
 class FrequencyByte{
 
     unsigned char element[256];
-    long frequency [256] = {0};
+    int frequency [256] = {0};
 
 public:
     FrequencyByte();
 
     char getElement(unsigned char element);
-    long getFrequency(unsigned char element);
+    int getFrequency(unsigned char element);
     void setFrequency(unsigned char frequency,int id);
     void plusFrequency(unsigned char element);
